refactor(chrome-extension): review component names
