+++ apps/chrome-extension/src/features/textProfileActiveDropdown/ui/TextProfileActiveDropdown.vue
