--- apps/chrome-extension/src/features/textProfileActiveEdit/index.ts
@@ -1,1 +1,0 @@
-export { default as TextProfileActiveEdit } from './ui/TextProfileActiveEdit.vue'
