--- apps/chrome-extension/src/features/textProfileSave/index.ts
@@ -1,1 +1,0 @@
-export { default as TextProfileSave } from './ui/TextProfileSave.vue'
