+++ apps/chrome-extension/src/features/textProfileSaveButton/index.ts
@@ -1,0 +1,1 @@
+export { default as TextProfileSaveButton } from './ui/TextProfileSaveButton.vue'
