+++ apps/chrome-extension/src/features/textProfileSaveButton/ui/TextProfileSaveButton.spec.ts
@@ -1,10 +1,10 @@
 import { mount } from '@vue/test-utils'
-import TextProfileSave from './TextProfileSave.vue'
+import TextProfileSave from './TextProfileSaveButton.vue'
 import { textSettingsFixture as settings, textProfileFixture as profile, useTextPreferences } from '@/entities/textPreferences'
 import { mockAlert, mockPrompt } from '@/shared/test'
 import { Settings } from '@readapt/settings'
 
-describe('TextProfileSave', () => {
+describe('TextProfileSaveButton', () => {
   afterEach(() => {
     useTextPreferences().reset()
     jest.restoreAllMocks()
